Upd test projects naming; Add wiki folder
