+++ tests/Structr.Tests.Abstractions/Structr.Tests.Abstractions.csproj
