+++ tests/Structr.Tests.Domain/EntityTests.cs
@@ -1,7 +1,7 @@
 using System;
 using Xunit;
 
-namespace Structr.Domain.Tests
+namespace Structr.Tests.Domain
 {
     public class EntityTests
     {
