+++ tests/Structr.Tests.Domain/Structr.Tests.Domain.csproj
