+++ tests/Structr.Tests.IO/Structr.Tests.IO.csproj
