+++ tests/Structr.Tests.Navigation/MenuItem.cs
@@ -1,4 +1,6 @@
-namespace Structr.Navigation.Tests
+using Structr.Navigation;
+
+namespace Structr.Tests.Navigation
 {
     public class MenuItem : NavigationItem<MenuItem>
     {
