+++ tests/Structr.Tests.Specifications/NoneSpecificationTests.cs
@@ -1,14 +1,15 @@
+using Structr.Specifications;
 using System.Collections.Generic;
 using Xunit;
 
-namespace Structr.Specifications.Tests
+namespace Structr.Tests.Specifications
 {
-    public class AnySpecificationTests
+    public class NoneSpecificationTests
     {
         [Fact]
-        public void IsSatisfiedBy_StringCollection_AllItems()
+        public void IsSatisfiedBy_StringCollection_NoneItems()
         {
-            var spec = new AnySpecification<string>();
+            var spec = new NoneSpecification<string>();
 
             List<string> actual = new List<string>();
             foreach (var item in StringCollection.Items)
@@ -19,7 +20,7 @@ namespace Structr.Specifications.Tests
                 }
             }
 
-            Assert.Equal(StringCollection.Items, actual);
+            Assert.Equal(StringCollection.Empty, actual);
         }
     }
 }
