+++ tests/Structr.Tests.Specifications/Structr.Tests.Specifications.csproj
